modify pipeline for "update_by_data"; fix bug in speciate. currently, node_delete and conn_delete can successfully work

+++ tensorneat/problem/func_fit/xor.py
@@ -4,14 +4,19 @@ from .func_fit import FuncFit
 
 
 class XOR(FuncFit):
-
     @property
     def inputs(self):
-        return np.array([[0, 0], [0, 1], [1, 0], [1, 1]])
+        return np.array(
+            [[0, 0], [0, 1], [1, 0], [1, 1]],
+            dtype=np.float32,
+        )
 
     @property
     def targets(self):
-        return np.array([[0], [1], [1], [0]])
+        return np.array(
+            [[0], [1], [1], [0]],
+            dtype=np.float32,
+        )
 
     @property
     def input_shape(self):
