+++ tensorneat/problem/func_fit/xor3d.py
@@ -16,12 +16,16 @@ class XOR3d(FuncFit):
                 [1, 0, 1],
                 [1, 1, 0],
                 [1, 1, 1],
-            ]
+            ],
+            dtype=np.float32,
         )
 
     @property
     def targets(self):
-        return np.array([[0], [1], [1], [0], [1], [0], [0], [1]])
+        return np.array(
+            [[0], [1], [1], [0], [1], [0], [0], [1]],
+            dtype=np.float32,
+        )
 
     @property
     def input_shape(self):
